Model and Cleaning Work
I added in multiple models which aren't working perfectly yet, but am going to clean them up, using all the stats to then perdict what is most important to affecting +/-. Also had to clean a bit of the data as anything that had a space, or sign like % or -,+,/ or any spaces wouldnt run within the model. I also removed the app from the qmd as we already have the app running on a seperate app doc.